+++ CHANGES.rst
@@ -17,6 +17,7 @@ Major Changes
 - build: add missing wheel dependency to dev installs
 - build: add missing setuptools dependency to dev installs
 - build: send build data over sockets to a local signing service
+- test: add full coverage for '_paths.py'
 - perf: store PackageSigner files as hexdigests to ease transmission
 - refactor: update '_paths.py' style, names & logic structure